Enable translation while displaying an alternative source language.
A user option is provided to choose which language should be used as the 
alternative source language. The feature can be enabled/disabled server wide
with the variable Pootle.enablealtsrc (disabled by default).

This closes bug 209.

Known issues:
Plural units are not supported yet (nothing is displayed)
Unit context (such as msgctxt in PO) is not taken into account
The first match from the alternative source language project is used - not 
specifically from the unit or file corresponding to the current unit
This might need lots of memory and possibly processing power in large projects

[Contributed by Julen Ruiz Aizpuru as part of GSoC2008, based on work by 
Sayamindu Dasgupta]


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/Pootle@7535 54714841-351b-0410-a198-e36a94b762f5