Fixed a mismatch in the way the quality checks looked at suggestions and in the way that Pootle looked at suggestions.
When the suggestion quality check looks at whether a unit has a
suggestion, it simply looks to see whether there exists a suggestion
with a matching source string.

In contrast, Pootle also required that the location information
of a unit in a suggestion file matched.

This caused problems if the location information in translation
files was updated. This is because the quality check mechanism
would report that there were suggestions, while Pootle would
fail to find the suggestions, due to the mismatch of location
information.

Pootle now also only looks at the source information. This is not
ideal and we need a more fundamental solution. But for now, this
is the best way to go.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/Pootle@8176 54714841-351b-0410-a198-e36a94b762f5