Although this patch is big, the changes are rather simple.
1. The templates were using a variable called request, which was not the
same as the request object used in the views. This request was renamed
to its old name, sessionvars.

2. sessionvars was computed in many places, where it only needs to be
computed once, in PootlePage; this was fixed.

3. If we want to use Django's AuthenticationForm, then we have to set a
test cookie which is used by Django to determine if the user's browser
has cookies enabled. We have such a form on each page, since the user
can log in from any page. Thus, we simply set the test cookie in
middleware (it is only set if the user is not logged in).

4. Changed the login form in pootlepage.kid to a Django form.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/Pootle@9643 54714841-351b-0410-a198-e36a94b762f5