Moved the rights to the database.
The Right model implements the rights that a user has for a
particular TranslationProject.

Now, when a TranslationProject object is created, we automatically
create rights for the default user and the nobody user. These users
are added in initdb.py.

Note that the interface to getrights has changed. It now takes a
user object.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/Pootle@9675 54714841-351b-0410-a198-e36a94b762f5