Improved the rights checks a bit.
The function redirect is a thin convience wrapper around Django's
HTTP redirect functionality. It takes all arguments after the
redirection url as HTTP GET arguments.

In the code, redirect is used to redirect the user if she does not
have appropriate rights to do something. The HTTP GET parameter
"message" is used to pass information about why the request was
rejected.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/Pootle@9687 54714841-351b-0410-a198-e36a94b762f5