The most notable change of this patch is that it replaces the direct URL creation with an abstraction that makes it appear as if we're calling functions. So, instead of builing a URL using the old URL class, we now have something like
    dispatch.show_directory

show_directory will create a link that will do the right thing
for us.

For each kind of page in Pootle, there's a certain amount of
state that must be carried along. This state isn't session
state, but is contained in the URLs. In order to deal with
this properly, there are now State subclasses, such as
ProjectIndexState (for project index pages) and TranslatePageState
(for translation pages). These can be constructed at any
time from GET or POST variables.

The functions that replace the URL creation (those in dispatch)
use state objects to ensure that proper transfer of state
from the current URL to the new URL constructed by the
function.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/Pootle@10599 54714841-351b-0410-a198-e36a94b762f5