Properly check for profile permissions.
An store string was being passed instead of an instance of the
Directory model.

Also, this now raises another problem: anonymous and default users may
have problems when submitting translations/suggestions for the first
time since we have the captcha middleware. This needs to be adapted to
understand xhr requests and act accordingly.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk/Pootle@15900 54714841-351b-0410-a198-e36a94b762f5